+++ src/components/ChangeHowTheySeeYouModal.js
@@ -1,4 +1,4 @@
-import React, { useState } from "react";
+import React, { useState, useLayoutEffect } from "react";
 import {
   Image,
   View,
@@ -10,6 +10,20 @@ import {
   Keyboard,
   Button,
 } from "react-native";
+import {
+  collection,
+  query,
+  where,
+  getDoc,
+  getDocs,
+  addDoc,
+  onSnapshot,
+  doc,
+  updateDoc,
+  serverTimestamp,
+  limit,
+} from "firebase/firestore";
+import { auth, database } from "../../config/firebase";
 import useIcon from "../hooks/useIcon";
 import { Overlay } from "react-native-elements";
 import modalStyles from "./utils/modalStyles";
@@ -18,6 +32,10 @@ import OSIOptionBox from "./OSIOptionBox";
 import DropDownPicker from "react-native-dropdown-picker";
 import DateTimePickerModal from "react-native-modal-datetime-picker";
 import ModalDropdown from "react-native-modal-dropdown";
+import useFetchBubbleMembers from "../hooks/useFetchBubbleMembers";
+import defaultStatusStyles from "../screens/Home/utils/defaultStatusStyles";
+import useIndicator from "../hooks/useIndicator";
+import RadioButton from "../components/RadioButton";
 
 const ChangeHowTheySeeYouModal = (props) => {
   const slideUpDownIcon = useIcon("slideUpDownIcon");
@@ -25,52 +43,38 @@ const ChangeHowTheySeeYouModal = (props) => {
   const toggleOnIcon = useIcon("toggleOnIcon");
 
   const [isToggledOn, setToggledOn] = useState(false);
+  const [bubble, setBubble] = useState(null);
+  const [option, setOption] = useState(null);
+  const [customMessageValue, onChangeText] = useState(null);
+  const [members, setMembers] = useState([]);
   const toggleIcon = isToggledOn ? toggleOnIcon : toggleOffIcon;
 
   const isModalVisible = props.is2ndModalVisible;
-  //const setModalVisible = props.set2ndModalVisible;
+  const setModalVisible = props.set2ndModalVisible;
 
-  const [isOSISelected, setIsOSISelected] = useState([
+  const radioData = [
     {
-      id: 1,
-      value: "openToChat",
-      indicator: "openToChat",
-      name: "Open To Chat",
-      selected: true,
+      id: "Open to Chat",
+      value: "Open to Chat",
+      indicator: useIndicator("openToChat"),
+      useIndicator,
     },
     {
-      id: 2,
-      value: "idle",
-      indicator: "idle",
-      name: "Be Right Back",
-      selected: false,
+      id: "Be Right Back",
+      value: "Be Right Back",
+      indicator: useIndicator("idle"),
     },
     {
-      id: 3,
-      value: "doNotDisturb",
-      indicator: "doNotDisturb",
-      name: "Do Not Disturb",
-      selected: false,
+      id: "Do Not Disturb",
+      value: "Do Not Disturb",
+      indicator: useIndicator("doNotDisturb"),
     },
     {
-      id: 4,
-      value: "invisible",
-      indicator: "invisible",
-      name: "Invisible",
-      selected: false,
+      id: "Invisible",
+      value: "Invisible",
+      indicator: useIndicator("invisible"),
     },
-  ]);
-
-  const onRadioBtnClick = (item) => {
-    let updatedState = isOSISelected.map((isSelectedItem) =>
-      isSelectedItem.id === item.id
-        ? { ...isSelectedItem, selected: true }
-        : { ...isSelectedItem, selected: false }
-    );
-    setIsOSISelected(updatedState);
-  };
-
-  const [customMessageValue, onChangeText] = React.useState("");
+  ];
 
   const [openDropdown, setOpenDropdown] = useState(false);
   const [dropdownValue, setDropdownValue] = useState("1 hour");
@@ -137,6 +141,136 @@ const ChangeHowTheySeeYouModal = (props) => {
     hideDatePicker();
   };
 
+  function updateUserStatusFunction() {
+    const updateStatus = async () => {
+      const userRef = doc(database, "users", auth.currentUser.uid);
+      const dataDefaultSnap = await getDoc(userRef);
+      const currentTime = serverTimestamp();
+      var bubbleRef = bubble;
+
+      //Bubble making if there is no bubble
+      if (!bubble) {
+        console.log("No bubble");
+        bubbleRef = await addDoc(collection(database, "bubbles"), {
+          computerGenerated: true,
+          creatorID: userRef,
+          lastChanged: currentTime,
+          statusID: dataDefaultSnap.data().statusID,
+          title: props.bubbleTitle,
+          conversationID: doc(database, "conversations", props.convoID),
+        });
+
+        console.log("bubbleref", bubbleRef);
+        members.map(async (member) => {
+          var memberRef = doc(database, "users", member.id);
+          await addDoc(collection(database, "bubble_members"), {
+            bubbleID: bubbleRef,
+            memberID: memberRef,
+          });
+        });
+
+        console.log("Members added!");
+
+        setBubble(bubbleRef);
+      }
+
+      const dataSnap = await getDoc(bubbleRef);
+
+      if (
+        JSON.stringify(dataDefaultSnap.data().statusID) ==
+        JSON.stringify(dataSnap.data().statusID)
+      ) {
+        console.log("Equal");
+        const dataOSISnap = await addDoc(
+          collection(database, "online_statuses"),
+          {
+            expiry: null,
+            message: customMessageValue,
+            osi: option,
+            toggleTime: false,
+          }
+        )
+          .then(async (docRef) => {
+            await updateDoc(bubbleRef, {
+              lastChanged: currentTime,
+              statusID: docRef,
+            }).catch((error) => console.error(error));
+
+            console.log("Added and updated");
+          })
+          .catch((error) => console.error(error));
+      } else {
+        const dataOSISnap = await updateDoc(dataSnap.data().statusID, {
+          osi: option,
+          message: customMessageValue,
+        })
+          .then(() => {
+            console.log("Status updated!");
+          })
+          .catch((error) => console.error(error));
+      }
+    };
+
+    updateStatus();
+  }
+
+  useLayoutEffect(() => {
+    const initialUpdate = async () => {
+      const userRef = doc(database, "users", auth.currentUser.uid);
+      var bubbleRef;
+      var q;
+
+      //Look for bubble, either through bubbleID or through convoID
+      console.log("There is no bubble: " + props.convoID);
+      const conversationRef = doc(database, "conversations", props.convoID);
+
+      q = query(
+        collection(database, "bubbles"),
+        where("creatorID", "==", userRef),
+        where("conversationID", "==", conversationRef),
+        limit(1)
+      );
+
+      const querySnapshot = await getDocs(q);
+
+      querySnapshot.forEach(async (convBubble) => {
+        console.log("There is a convo bubble", convBubble.id);
+        bubbleRef = doc(database, "bubbles", convBubble.id);
+
+        setBubble(bubbleRef);
+        const docSnap = await getDoc(bubbleRef);
+        //Get members
+        const bubbleMembers = await useFetchBubbleMembers(docSnap.id);
+        setMembers(bubbleMembers);
+
+        //Set default options
+        const dataOSISnap = await getDoc(docSnap.data().statusID);
+        setOption(dataOSISnap.data().osi);
+        onChangeText(dataOSISnap.data().message);
+      });
+
+      //If there are no bubbles for the conversation
+      if (querySnapshot.empty) {
+        console.log("There is no bubble at all");
+        const docSnap = await getDoc(userRef);
+
+        //Get members
+        const conversationMembers = await useFetchConversationUsers(
+          props.convoID,
+          false
+        );
+        setMembers(conversationMembers);
+
+        //Set default options
+        const dataOSISnap = await getDoc(docSnap.data().statusID);
+        setOption(dataOSISnap.data().osi);
+        onChangeText(dataOSISnap.data().message);
+      }
+    };
+
+    initialUpdate();
+  }, []);
+
   return (
     <View style={modalStyles.modalStatusContainer}>
       <GestureRecognizer onSwipeDown={() => props.set1stModalVisible(false)}>
@@ -171,19 +305,25 @@ const ChangeHowTheySeeYouModal = (props) => {
             <View>
               <Text style={modalStyles.modalHeaderText}>Your Status</Text>
               <Text style={modalStyles.modalHeaderSubtext}>
-                for FirstName LastName
+                for {props.title}
               </Text>
 
-              {isOSISelected.map((item) => (
-                <OSIOptionBox
-                  value={item.value}
-                  indicator={item.indicator}
-                  name={item.name}
-                  selected={item.selected}
-                  key={item.id}
-                  onPress={() => onRadioBtnClick(item)}
+              <View style={modalStyles.optionOSIContainer}>
+                <RadioButton
+                  key={radioData}
+                  data={radioData}
+                  onSelect={(value) => setOption(value)}
+                  current={option}
                 />
-              ))}
+              </View>
+              {/* <OSIOptionBox
+                value={item.value}
+                indicator={item.indicator}
+                name={item.name}
+                selected={item.selected}
+                key={item.id}
+                onPress={() => onRadioBtnClick(item)}
+              /> */}
 
               <View>
                 <View style={modalStyles.modalSubheaderTextContainer}>
@@ -244,7 +384,8 @@ const ChangeHowTheySeeYouModal = (props) => {
                 </View>
               ) : null}
 
-              <TouchableOpacity onPress={() => setToggledOn(!isToggledOn)}>
+              {/* TO DO display to others the clear time */}
+              {/* <TouchableOpacity onPress={() => setToggledOn(!isToggledOn)}>
                 <View style={modalStyles.modalSubheaderTextContainer}>
                   <Text style={modalStyles.modalSubheaderSubtext}>
                     Display duration of status to others?
@@ -254,10 +395,13 @@ const ChangeHowTheySeeYouModal = (props) => {
                     style={modalStyles.modalToggleIcon}
                   />
                 </View>
-              </TouchableOpacity>
+              </TouchableOpacity> */}
 
               <TouchableOpacity
-                onPress={() => props.set1stModalVisible(false)}
+                onPress={() => {
+                  updateUserStatusFunction();
+                  props.set1stModalVisible(false);
+                }}
                 activeOpacity={0.5}
                 style={modalStyles.modalSaveButtonContainer}
               >